优化修正日志管理器代码
