+++ app/src/main/kotlin/indi/dmzz_yyhyy/lightnovelreader/data/logging/LogEntry.kt
@@ -1,0 +1,6 @@
+package indi.dmzz_yyhyy.lightnovelreader.data.logging
+
+data class LogEntry(
+    val text: String,
+    val logLevel: LogLevel
+)
